Fix the invalid prober version issue
